bash: Enable smartcase searching for less(1)
I usually do not care about casing when I am searching for a term. In
fact, having letter casing be a factor when searching in less(1) can
make it harder for me to find what I am looking for. There are
situations where I do care about casing though, so a nice solution is to
use smartcasing; ignore casing unless a capital is found within the
search term.