vim: Display value of hlsearch when toggling
When toggling hlsearch using my mapping, the value shown in the vim
commandline is just 'hlsearch!'. This gives little indication to the
current value of hlsearch.

This commit changes the output of the mapping to 'hlsearch' if hlsearch
is enabled and 'nohlsearch' if hlsearch is disabled, which is much more
clear.