vim: Always show invisible chars
I previously had a list of characters to use in place of invisible
characters and then toggled list to either show or hide these
characters (based on the previous value of list).

I almost always want to see invisible characters, so this commit always
shows invisible characters instead of toggling them. Doesn't really
change much to be honest, but it makes things a little more clear.