vim: Adjust line number colours
The slate colorscheme has different colours for the current line number
and the other line numbers when using hybrid line numbers. The pablo
colorscheme does not have this by default, but it looks kind of cool.

This commit adjusts my custom pablo colorscheme (my-pablo) so that the
curent line number has a different styling than the other line numbers.