vim: Revert line number formatting
This partially reverts the commit that adjusted styling of the line
numbers in hybrid line number mode. I decided that I would rather have
the current line number and other line numbers have the same formatting
(like before). It is still easy to differentiate the cursor line number
from the other line numbers as the cursor line number is leff-justified
and the other line numbers are right-justified.