fzf: Tweak fzf current line bg colour
In some terminal emulators (like gnome-terminal + system colour theme),
the contrast between the terminal's background colour and fzf's current
line background colour is much too low.

This commit alters the colour of the curent line's background so that it
is actually visible and distinct in gnome-terminal (with the default
colour scheme).