ALL: Use stow's 'dot-' preprocessing for dotfiles
This makes it easier to work with dotfiles as they are no longer hidden.
A .stowrc file is used for convenience (contains the --dotfile option so
that it doesn't need to be specified for each invocation of stow).

Unfortunately, this does not seem to work for cases where the directory
already exists. Example: trying to use dot-config/ instead of .config/
in gtk/ will raise an error if ~/.config/ exists. gtk/ is left untouched
for now.

Specifying filetypes with vim modelines is needed by some files since
'dot-' preprocessing changes the names of some files e.g. .bashrc is
recognized as an sh file while dot-bashrc is not.