gtk,gvim: Use Terminus OTB instead of TTF
The TrueType version of the Terminus font was obtained from the AUR.
However, an OpenType version of Terminus was made available in Arch
Linux's community repo around the start of August. OpenType bitmap fonts
are also supported by pango and harfbuzz which means I don't really need
terminus-font-ttf anymore. terminus-font-otb also seems to stay more
up-to-date with terminus-font, so it would be better for me to just
switch from terminus-font-ttf to terminus-font-otb. terminus-font-otb is
also provided by the official Arch repos as opposed to terminus-font-ttf
which is another big plus for terminus-font-otb.