bash: Change venv formatting in bash prompt
The previous way of displaying python virtualenv information in the
prompt was satisfactory. I have decided to move venv info from the
beginning of the prompt to after the directory path (before git
information) to make the prompt look better. I have also added an arrow
that before the venv to differentiate it from the dir path.