bash: Add notes about hacky history functions
I wanted some more functionality with bash's history, similar to that
provided by zsh, so I wrote some hacky things to do that. Does not work
well with heredocs because heredocs span multiple lines in the history
file. zsh deals with this better, but I am stuck on bash due to command
line completions (some are not so easy to port to zsh).