ssh: Don't use dot- prefix
As of Stow 2.3.1, the dot- prefix option for config files is kind of
broken. Trying to stow a prefixed directory when the dot directory
already exists causes Stow to fail. This should be addressed sometime in
the near future, but for now, it would be better to rename ssh/dot-ssh
to ssh/.ssh to avoid this problem.

It's probably better for ~/.ssh to be a directory rather than a symlink
so that the permissions of ~/.ssh don't need to be changed everytime
when stowing on a new system.