r: Dynamically set width
The width when printing vectors, matrices, arrays, etc. is typically
fixed to a default width of 80 columns.

This can be troublesome when the R window is re-sized as lines may start
to wrap if the width is too large. On the other hand, if the width is
much smaller than the window's width, then not all of the screen will be
used, leading to inefficient use of screen real estate. A fix for this
is to dynamically set the initial width and let R change the width
whenever the R window is re-sized.