bash: Localize variables using fewer `local`s
Previously, each local variable in a function in ~/.bashrc is declared
using a separate `local`. However, multiple variables can actually be
localized with a single `local`. Related variables like `venv` and
`venv_prefix` should be declared local with a single `local` to reduce
the number of lines and to increase code readability.