bash: Fix HISTFILE truncation by invocation of sh
sh is a symlink to bash on Arch Linux. Both shells will share the same
HISTFILE (by default), but they will ont share the same HISTSIZE or
HISTFILESIZE unless explicitly set. Ultimately, this means that HISTFILE
is at risk of truncation if HISTFILE and/or HISTFILESIZE are set to
values larger than 500 in ~/.bashrc.

An easy fix is to have sh use the same HISTFILE and HISTFILESIZE values
as bash by exporting the aforementioned variables in ~/.bashrc.