gtk-theme: Add "Adwaita-Short-Panel" theme
GNOME 40 increases the height of the panel from 1.86em to 2.2em.
The "Adwaita-Short-Panel" theme reverts this change.

The change in panel height may be undesirable to some users due to the
way that it affects text height alignment in GNOME Terminal.

Settings related to the panel cannot be changed in a GTK CSS file, but
these settings can be changed in a separate theme. The
"Adwaita-Short-Panel" theme reverts the panel height by only changing
this component of GNOME Shell. Selecting this theme through GNOME Tweaks
will revert the panel height back to 1.86em.