gtk: Remove Terminus TTF-related information
terminus-fonts 4.49-2 on Arch Linux ships OTB fonts (called Terminus),
which as mentioned in the gtk.css comments, work fine with harfbuzz and
pango (and thus gnome-terminal). There is no longer a reason to use
terminus-fonts-ttf from the AUR, so remove mentions of it.

Also shorten Stack Overflow URL so that its not overly long.