readme: Link to xero's GitHub mirror
blog.xero.nu appears to be down as of writing, so use a link that isn't.
The linked README appears to be mostly the same as the old blog post
anyways, which is nice.

Also link to more specific section in README since it covers more than
just using stow.