gtk,gtk-themes: Give up on terminal font alignment
The purpose of the gtk.css and gnome-shell.css files were to adjust the
height of the gnome-terminal window (when maximized) so that using the
Terminus font (size 13) would result in pixel-perfect alignment with the
bottom of an 1080p screen.

However, the pango 1:1.50.0-1 release on Arch Linux seems to have
increased vertical font spacing resulting in a small gap between the
last line in gnome-terminal and the bottom of the screen.

Pixel-perfect alignment can still be achieved with the previous CSS
files and with Terminus 12 and 1x1.1 cell spacing, but this results in
text that is just a little too small for comfort on a 14-inch screen.

As of writing, the gap resulting from this change (with Terminus 13)
isn't actually all that noticeable (see [1]) and it's a huge pain to
achieve pixel-perfect alignment without sacrificing readability, so I'm
giving up on it (for now) and deleting these two CSS files since they no
longer serve their original purpose.

[1] https://i.imgur.com/oXdCgYI.png